Split tests out into two files
