+++ tests/src/Unit/Views/ViewEventTest.php
@@ -3,17 +3,15 @@
 namespace Drupal\Tests\hook_event_dispatcher\Unit\Views;
 
 use Drupal\Core\DependencyInjection\ContainerBuilder;
-use Drupal\hook_event_dispatcher\Event\Views\ViewsDataAlterEvent;
-use Drupal\hook_event_dispatcher\Event\Views\ViewsDataEvent;
 use Drupal\hook_event_dispatcher\Event\Views\ViewsPostRenderEvent;
-use Drupal\hook_event_dispatcher\Event\Views\ViewsPreRenderEvent;
 use Drupal\hook_event_dispatcher\Event\Views\ViewsPreViewEvent;
+use Drupal\hook_event_dispatcher\Event\Views\ViewsQuerySubstitutionsEvent;
 use Drupal\hook_event_dispatcher\HookEventDispatcherInterface;
 use Drupal\Tests\hook_event_dispatcher\Unit\HookEventDispatcherManagerSpy;
+use Drupal\Tests\UnitTestCase;
 use Drupal\views\Plugin\views\cache\CachePluginBase;
 use Drupal\views\Plugin\views\query\QueryPluginBase;
 use Drupal\views\ViewExecutable;
-use Drupal\Tests\UnitTestCase;
 
 /**
  * Class ViewEventTest.
@@ -43,123 +41,13 @@ class ViewEventTest extends UnitTestCase {
   }
 
   /**
-   * ViewsDataEvent test.
-   */
-  public function testViewsDataEvent() {
-    $data = [
-      'test' => [
-        'test_array_data',
-      ],
-    ];
-
-    $this->manager->setEventCallbacks([
-      HookEventDispatcherInterface::VIEWS_DATA => function (ViewsDataEvent $event) use ($data) {
-        $event->addData($data);
-      },
-    ]);
-
-    $result = hook_event_dispatcher_views_data();
-
-    $this->assertSame($data, $result);
-  }
-
-  /**
-   * ViewsDataEvent multiple adds test.
-   */
-  public function testViewsDataEventMultipleAdds() {
-    $data1 = [
-      'test' => [
-        'test_array_data',
-      ],
-    ];
-
-    $data2 = [
-      'test' => [
-        'other_test_array_data',
-      ],
-    ];
-
-    $data3 = [
-      'some' => [
-        'other_data',
-      ],
-    ];
-
-    $this->manager->setEventCallbacks([
-      HookEventDispatcherInterface::VIEWS_DATA => function (ViewsDataEvent $event) use ($data1, $data2, $data3) {
-        $event->addData($data1);
-        $event->addData($data2);
-        $event->addData($data3);
-      },
-    ]);
-
-    $result = hook_event_dispatcher_views_data();
-
-    $expectedResult = [
-      'test' => [
-        'test_array_data',
-        'other_test_array_data',
-      ],
-      'some' => [
-        'other_data',
-      ],
-    ];
-    $this->assertSame($expectedResult, $result);
-  }
-
-  /**
-   * ViewsDataAlterEvent by reference test.
-   */
-  public function testViewsDataAlterEventByReference() {
-    $this->manager->setEventCallbacks([
-      HookEventDispatcherInterface::VIEWS_DATA_ALTER => function (ViewsDataAlterEvent $event) {
-        $data = &$event->getData();
-        $data['test']['other_test'] = ['some_data'];
-      },
-    ]);
-
-    $data = $expectedData = [
-      'test' => [
-        'test' => 'test_array_data',
-      ],
-    ];
-    hook_event_dispatcher_views_data_alter($data);
-
-    $expectedData['test']['other_test'] = ['some_data'];
-    $this->assertSame($expectedData, $data);
-  }
-
-  /**
-   * ViewsDataAlterEvent by set test.
-   */
-  public function testViewsDataAlterEventBySet() {
-    $this->manager->setEventCallbacks([
-      HookEventDispatcherInterface::VIEWS_DATA_ALTER => function (ViewsDataAlterEvent $event) {
-        $data = $event->getData();
-        $data['other'] = ['other_data'];
-        $event->setData($data);
-      },
-    ]);
-
-    $data = $expectedData = [
-      'test' => [
-        'test' => 'test_array_data',
-      ],
-    ];
-    hook_event_dispatcher_views_data_alter($data);
-
-    $expectedData['other'] = ['other_data'];
-    $this->assertSame($expectedData, $data);
-  }
-
-  /**
    * Pre view event.
    */
   public function testPreViewEvent() {
     /** @var \Drupal\views\ViewExecutable $view */
     $view = $this->createMock(ViewExecutable::class);
-    $display_id = 'test';
-    $args = $original_args = [
+    $displayId = 'test';
+    $args = [
       'test',
     ];
 
@@ -170,7 +58,7 @@ class ViewEventTest extends UnitTestCase {
       },
     ]);
 
-    hook_event_dispatcher_views_pre_view($view, $display_id, $args);
+    hook_event_dispatcher_views_pre_view($view, $displayId, $args);
 
     $this->assertEquals('test2', $args[0]);
   }
@@ -294,12 +182,16 @@ class ViewEventTest extends UnitTestCase {
   public function testQuerySubstitions() {
     /** @var \Drupal\views\ViewExecutable $view */
     $view = $this->createMock(ViewExecutable::class);
+    $expected_substitutions = array("test" => 1);
 
-    hook_event_dispatcher_views_query_substitutions($view);
+    $this->manager->setEventCallbacks([
+      HookEventDispatcherInterface::VIEWS_QUERY_SUBSTITUTIONS => function (ViewsQuerySubstitutionsEvent $event) use ($expected_substitutions) {
+        $event->setSubstitutions($expected_substitutions);
+      },
+    ]);
 
-    /* @var \Drupal\hook_event_dispatcher\Event\Views\ViewsQuerySubstitutionsEvent $event */
-    $event = $this->manager->getRegisteredEvent(HookEventDispatcherInterface::VIEWS_QUERY_SUBSTITUTIONS);
-    $this->assertSame($view, $event->getView());
+    $result = hook_event_dispatcher_views_query_substitutions($view);
+    $this->assertSame($expected_substitutions, $result);
   }
 
 }
